backend/requirements/base.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
django-filter==25.1
django-cors-headers==4.7.0
django-unfold==0.68.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.29.0
psycopg2-binary==2.9.11
python-decouple==3.8
PyJWT==2.9.0
PyYAML==6.0.3
tzdata==2025.1
sqlparse==0.5.3
uritemplate==4.2.0
pillow==11.2.1